Rename the data_sources module to content
