+++ tests/test_multiple_users.py
@@ -5,7 +5,7 @@ from urllib.parse import quote
 import pytest
 
 from khoj.database.models import KhojApiUser, KhojUser
-from khoj.processor.data_sources.org_mode.org_to_entries import OrgToEntries
+from khoj.processor.content.org_mode.org_to_entries import OrgToEntries
 
 # Internal Packages
 from khoj.search_type import text_search
