+++ tests/test_text_search.py
@@ -8,8 +8,8 @@ from pathlib import Path
 import pytest
 
 from khoj.database.models import Entry, GithubConfig, KhojUser, LocalOrgConfig
-from khoj.processor.data_sources.github.github_to_entries import GithubToEntries
-from khoj.processor.data_sources.org_mode.org_to_entries import OrgToEntries
+from khoj.processor.content.github.github_to_entries import GithubToEntries
+from khoj.processor.content.org_mode.org_to_entries import OrgToEntries
 
 # Internal Packages
 from khoj.search_type import text_search
